traffic-dashboard/backend/requirements.txt
fastapi[standard]>=0.104.1
uvicorn[standard]>=0.23.2
websockets>=11.0.3
eclipse-sumo>=1.19.0
traci>=1.19.0
torch>=2.0.0
numpy>=1.24.0
motor>=3.3.0
python-dotenv>=1.0.0
ultralytics>=8.1.0
yt-dlp>=2024.3.10
groq>=0.5.0
opencv-python>=4.9.0
matplotlib>=3.8.0
pandas>=2.2.0
